third_party/WebKit/LayoutTests/pointer-lock/pointer-lock-api.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body>
<script>
    description("Basic API existence test for Pointer Lock.")

    shouldBeDefined("document.onwebkitpointerlockchange");
    shouldBeDefined("document.onwebkitpointerlockerror");
    shouldBeDefined("document.webkitExitPointerLock");
    shouldEvaluateTo("document.webkitPointerLockElement", null);
    element = document.createElement('x');
    shouldBeDefined("element.webkitRequestPointerLock");
</script>
</body>
</html>
